task09-multimodal-video-analysis/requirements.txt
websockets
google-genai
python-dotenv
opencv-python  
pyaudio  
pillow  
mss